backend/requirements.txt
django==5.1.6
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
requests==2.31.0
matplotlib==3.8.2
gunicorn==23.0.0
whitenoise==6.8.2
djangorestframework-simplejwt==5.3.1

# Database Support for Render
dj-database-url==2.1.0
psycopg2-binary==2.9.9

# Authentication & Google Login
dj-rest-auth==6.0.0
django-allauth==0.61.1
cryptography==42.0.5

# AI Nutrition Estimation
google-generativeai>=0.8.0